unify types and set them all to int64

+++ onnx_tf/backends/backend_v5.py
@@ -19,13 +19,13 @@ class TensorflowBackend(TensorflowBackendBase):
   @classmethod
   def handle_reshape(cls, node, input_dict):
     tensor = input_dict[node.inputs[0]]
-    shape = input_dict[node.inputs[1]]
-    input_shape = tf.shape(tensor, out_type=tf.int32)
+    shape = tf.cast(input_dict[node.inputs[1]], tf.int64)
+    input_shape = tf.shape(tensor, out_type=tf.int64)
 
     # Extract indicies of the shape paramter where
     # a copy from the original dimension size is needed.
     copy_indices = tf.squeeze(tf.where(tf.equal(shape,
-                                                tf.constant(0, dtype=tf.int32))), -1)
+                                                tf.constant(0, dtype=tf.int64))), -1)
 
     indices_gathered = tf.gather(input_shape, copy_indices)
     indices_scattered = tf.sparse_to_dense(copy_indices,
